requirements.txt
asgiref==3.4.1
Django~=3.2.6
MarkupSafe==2.0.1
numpy~=1.21.1
pandas~=1.3.1
psycopg2==2.7.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
openpyxl~=3.0.7
matplotlib~=3.4.3
djangorestframework~=3.12.4
django-jsonfield~=1.4.1